+++ frontends/web/src/routes/lightning/lightning.tsx
@@ -21,7 +21,6 @@ import { GuideWrapper, GuidedContent, Header, Main } from '../../components/layo
 import { Spinner } from '../../components/spinner/Spinner';
 import { ActionButtons } from './components/action-buttons';
 import { LightningGuide } from './guide';
-import { LightningTorProxyWarning } from '@/components/banners/lightning-tor-proxy-warning';
 import { Status } from '../../components/status/status';
 import { HideAmountsButton } from '../../components/hideamountsbutton/hideamountsbutton';
 import { PaymentDetails } from './components/payment-details';
@@ -376,7 +375,6 @@ export const Lightning = () => {
 
   const statusBanners = (
     <>
-      <LightningTorProxyWarning dismissible />
       <Status
         dismissibleKey="lightning-beta-warning"
         type="warning">